refactor: isolate action logging and HTTP helpers
