+++ action/__init__.py
@@ -1,0 +1,1 @@
+"""Support utilities for the setup-wasi-sdk action."""
