+++ tests/test_install.py
@@ -8,6 +8,7 @@ from unittest import mock
 from urllib import error
 
 import install
+from action import http
 
 
 def make_archive(member_name, contents=b'contents', mode=0o644):
@@ -67,7 +68,7 @@ class ExtractArchiveTests(unittest.TestCase):
 
 class DownloadWithRetriesTests(unittest.TestCase):
     def setUp(self):
-        jitter = mock.patch('install.random.uniform', return_value=0)
+        jitter = mock.patch('action.http.random.uniform', return_value=0)
         self.random_uniform = jitter.start()
         self.addCleanup(jitter.stop)
 
@@ -76,8 +77,8 @@ class DownloadWithRetriesTests(unittest.TestCase):
         return error.HTTPError(
             'https://example.com/sdk.tar.gz', status, 'Error', headers or {}, None)
 
-    @mock.patch('install.time.sleep')
-    @mock.patch('install.request.urlretrieve')
+    @mock.patch('action.http.time.sleep')
+    @mock.patch('action.http.request.urlretrieve')
     def test_retries_network_errors_with_exponential_backoff(self, retrieve, sleep):
         retrieve.side_effect = [
             error.URLError(ConnectionResetError('connection reset')),
@@ -85,62 +86,62 @@ class DownloadWithRetriesTests(unittest.TestCase):
             None,
         ]
 
-        install.download_with_retries('https://example.com/sdk.tar.gz', '/tmp/sdk.tar.gz')
+        http.download_with_retries('https://example.com/sdk.tar.gz', '/tmp/sdk.tar.gz')
 
         self.assertEqual(retrieve.call_count, 3)
         sleep.assert_has_calls([mock.call(1), mock.call(2)])
 
-    @mock.patch('install.time.sleep')
-    @mock.patch('install.request.urlretrieve')
+    @mock.patch('action.http.time.sleep')
+    @mock.patch('action.http.request.urlretrieve')
     def test_retries_transient_http_errors(self, retrieve, sleep):
         unavailable = self.http_error(503)
         self.addCleanup(unavailable.close)
         retrieve.side_effect = [unavailable, None]
 
-        install.download_with_retries('https://example.com/sdk.tar.gz', '/tmp/sdk.tar.gz')
+        http.download_with_retries('https://example.com/sdk.tar.gz', '/tmp/sdk.tar.gz')
 
         self.assertEqual(retrieve.call_count, 2)
         sleep.assert_called_once_with(1)
 
-    @mock.patch('install.time.sleep')
-    @mock.patch('install.request.urlretrieve')
+    @mock.patch('action.http.time.sleep')
+    @mock.patch('action.http.request.urlretrieve')
     def test_does_not_retry_permanent_http_errors(self, retrieve, sleep):
         not_found = self.http_error(404)
         self.addCleanup(not_found.close)
         retrieve.side_effect = not_found
 
         with self.assertRaises(error.HTTPError):
-            install.download_with_retries(
+            http.download_with_retries(
                 'https://example.com/sdk.tar.gz', '/tmp/sdk.tar.gz')
 
         retrieve.assert_called_once()
         sleep.assert_not_called()
 
-    @mock.patch('install.time.sleep')
-    @mock.patch('install.request.urlretrieve')
+    @mock.patch('action.http.time.sleep')
+    @mock.patch('action.http.request.urlretrieve')
     def test_raises_after_all_attempts(self, retrieve, sleep):
         retrieve.side_effect = error.URLError('network unavailable')
 
         with self.assertRaises(error.URLError):
-            install.download_with_retries(
+            http.download_with_retries(
                 'https://example.com/sdk.tar.gz', '/tmp/sdk.tar.gz')
 
-        self.assertEqual(retrieve.call_count, install.DOWNLOAD_ATTEMPTS)
+        self.assertEqual(retrieve.call_count, http.DOWNLOAD_ATTEMPTS)
         sleep.assert_has_calls([mock.call(1), mock.call(2), mock.call(4)])
 
-    @mock.patch('install.time.sleep')
-    @mock.patch('install.request.urlretrieve')
+    @mock.patch('action.http.time.sleep')
+    @mock.patch('action.http.request.urlretrieve')
     def test_adds_bounded_jitter_to_backoff(self, retrieve, sleep):
         self.random_uniform.return_value = 0.75
         retrieve.side_effect = [error.URLError('network unavailable'), None]
 
-        install.download_with_retries('https://example.com/sdk.tar.gz', '/tmp/sdk.tar.gz')
+        http.download_with_retries('https://example.com/sdk.tar.gz', '/tmp/sdk.tar.gz')
 
-        self.random_uniform.assert_called_once_with(0, install.DOWNLOAD_JITTER_SECONDS)
+        self.random_uniform.assert_called_once_with(0, http.DOWNLOAD_JITTER_SECONDS)
         sleep.assert_called_once_with(1.75)
 
-    @mock.patch('install.time.sleep')
-    @mock.patch('install.request.urlretrieve')
+    @mock.patch('action.http.time.sleep')
+    @mock.patch('action.http.request.urlretrieve')
     def test_respects_retry_after(self, retrieve, sleep):
         headers = Message()
         headers['Retry-After'] = '17'
@@ -148,13 +149,13 @@ class DownloadWithRetriesTests(unittest.TestCase):
         self.addCleanup(rate_limited.close)
         retrieve.side_effect = [rate_limited, None]
 
-        install.download_with_retries('https://example.com/sdk.tar.gz', '/tmp/sdk.tar.gz')
+        http.download_with_retries('https://example.com/sdk.tar.gz', '/tmp/sdk.tar.gz')
 
         sleep.assert_called_once_with(17)
 
-    @mock.patch('install.time.time', return_value=1_000)
-    @mock.patch('install.time.sleep')
-    @mock.patch('install.request.urlretrieve')
+    @mock.patch('action.http.time.time', return_value=1_000)
+    @mock.patch('action.http.time.sleep')
+    @mock.patch('action.http.request.urlretrieve')
     def test_respects_github_rate_limit_reset(self, retrieve, sleep, _time):
         headers = Message()
         headers['X-RateLimit-Remaining'] = '0'
@@ -163,26 +164,26 @@ class DownloadWithRetriesTests(unittest.TestCase):
         self.addCleanup(rate_limited.close)
         retrieve.side_effect = [rate_limited, None]
 
-        install.download_with_retries('https://example.com/sdk.tar.gz', '/tmp/sdk.tar.gz')
+        http.download_with_retries('https://example.com/sdk.tar.gz', '/tmp/sdk.tar.gz')
 
         sleep.assert_called_once_with(42)
 
-    @mock.patch('install.time.sleep')
-    @mock.patch('install.request.urlretrieve')
+    @mock.patch('action.http.time.sleep')
+    @mock.patch('action.http.request.urlretrieve')
     def test_does_not_retry_unrelated_forbidden_response(self, retrieve, sleep):
         forbidden = self.http_error(403)
         self.addCleanup(forbidden.close)
         retrieve.side_effect = forbidden
 
         with self.assertRaises(error.HTTPError):
-            install.download_with_retries(
+            http.download_with_retries(
                 'https://example.com/sdk.tar.gz', '/tmp/sdk.tar.gz')
 
         retrieve.assert_called_once()
         sleep.assert_not_called()
 
-    @mock.patch('install.time.sleep')
-    @mock.patch('install.request.urlretrieve')
+    @mock.patch('action.http.time.sleep')
+    @mock.patch('action.http.request.urlretrieve')
     def test_does_not_retry_rate_limit_without_delay(self, retrieve, sleep):
         rate_limited = self.http_error(429)
         self.addCleanup(rate_limited.close)
@@ -190,25 +191,25 @@ class DownloadWithRetriesTests(unittest.TestCase):
 
         with self.assertLogs(level='ERROR') as logs:
             with self.assertRaises(error.HTTPError):
-                install.download_with_retries(
+                http.download_with_retries(
                     'https://example.com/sdk.tar.gz', '/tmp/sdk.tar.gz')
 
         self.assertIn('did not provide a usable retry delay', logs.output[0])
         retrieve.assert_called_once()
         sleep.assert_not_called()
 
-    @mock.patch('install.time.sleep')
-    @mock.patch('install.request.urlretrieve')
+    @mock.patch('action.http.time.sleep')
+    @mock.patch('action.http.request.urlretrieve')
     def test_does_not_wait_at_rate_limit_threshold(self, retrieve, sleep):
         headers = Message()
-        headers['Retry-After'] = str(install.MAX_RATE_LIMIT_WAIT_SECONDS)
+        headers['Retry-After'] = str(http.MAX_RATE_LIMIT_WAIT_SECONDS)
         rate_limited = self.http_error(429, headers)
         self.addCleanup(rate_limited.close)
         retrieve.side_effect = rate_limited
 
         with self.assertLogs(level='ERROR') as logs:
             with self.assertRaises(error.HTTPError):
-                install.download_with_retries(
+                http.download_with_retries(
                     'https://example.com/sdk.tar.gz', '/tmp/sdk.tar.gz')
 
         self.assertIn('is not less than', logs.output[0])
